requirements.txt
more_itertools==9.1.0
networkx==2.5.1
pydantic==1.10.2
PyYAML==6.0
setuptools==60.2.0
